Bug fix
Fixed a bug where configuring the fire button as either mouse button
would register the mouse click as '-' for the left button and '=' for
the right button. The player also wasn't able to use the mouse even if
the configuration code was working properly as it wasn't implemented in
the game code.